REFACTOR - Restructured Frontend
