+++ frontend/src/App.jsx
@@ -1,20 +1,20 @@
 import { motion } from 'framer-motion';
 import { useState, useEffect, useCallback, useRef } from 'react';
 import { Sun, Moon } from 'lucide-react';
-import Sidebar from './components/Sidebar';
-import SearchBar from './components/SearchBar';
-import PhotoGrid from './components/ImageGrid';
-import VideoGrid from './components/VideoGrid';
-import RawImageGrid from './components/RawImageGrid';
-import AllMediaGrid from './components/AllMediaGrid';
-import FavoritesGrid from './components/FavoritesGrid';
-import MapView from './components/MapView';
-import Albums from './components/Albums';
-import ProgressBar from './components/ProgressBar';
-import ComingSoon from './components/ComingSoon';
-import Dashboard from './components/Dashboard';
-import UploadButton from './components/UploadButton';
-import InsufficientSpaceModal from './components/InsufficientSpaceModal';
+import Sidebar from './components/layout/Sidebar';
+import SearchBar from './components/common/SearchBar';
+import PhotoGrid from './components/screens/PhotoGrid';
+import VideoGrid from './components/screens/VideoGrid';
+import RawImageGrid from './components/screens/RawImageGrid';
+import AllMediaGrid from './components/screens/AllMediaGrid';
+import FavoritesGrid from './components/screens/FavoritesGrid';
+import MapView from './components/screens/MapView';
+import Albums from './components/screens/Albums';
+import ProgressBar from './components/common/ProgressBar';
+import ComingSoon from './components/modals/ComingSoonModal';
+import Dashboard from './components/screens/Dashboard';
+import UploadButton from './components/buttons/UploadButton';
+import InsufficientSpaceModal from './components/modals/InsufficientSpaceModal';
 import { useTheme } from './contexts/ThemeContext';
 
 import { api } from './api';
