+++ frontend/src/apis/raw_images.js
