+++ frontend/src/components/buttons/DownloadButton.jsx
@@ -1,6 +1,6 @@
 import { Download, Loader2 } from 'lucide-react';
 import { useState } from 'react';
-import { api } from '../api';
+import { api } from '../../api';
 
 /**
  * Reusable Download Button Component
