+++ frontend/src/components/common/ProgressBar.jsx
