+++ frontend/src/components/common/SearchBar.jsx
