+++ frontend/src/components/modals/DeleteAlbumModal.jsx
@@ -1,7 +1,7 @@
 import { motion, AnimatePresence } from "framer-motion";
 import { useState, useEffect } from "react";
 import { X, AlertTriangle } from "lucide-react";
-import { api } from "../api";
+import { api } from "../../api";
 
 export default function DeleteAlbumModal({ isOpen, onClose, onSuccess, albumId, albumName, updateProgressBar, removeProgressBar, hasActiveDownload, cancelAlbumDownload }) {
   const [error, setError] = useState(null);
