+++ frontend/src/components/screens/AlbumGrid.jsx
@@ -2,24 +2,24 @@ import { motion, AnimatePresence } from "framer-motion";
 import { Play, Pause } from "lucide-react";
 import { useState, useMemo, useRef, useEffect, useCallback } from "react";
 import { Virtuoso } from "react-virtuoso";
-import { processTimelineData } from "../utils/timelineUtils";
-import ImageViewer from "./ImageViewer";
-import VideoViewer from "./VideoViewer";
-import RawImageViewer from "./RawImageViewer";
-import AddToAlbumModal from "./AddToAlbumModal";
-import { api } from "../api";
-import formatTime from "../utils/timeUtils";
-import { formatFileSize } from "../utils/fileUtils";
-import FavoriteButton from "./FavoriteButton";
-import ShareButton from "./ShareButton";
-import DownloadButton from "./DownloadButton";
-import DeleteButton from "./DeleteButton";
-import EditAlbumModal from "./EditAlbumModal";
-import DeleteAlbumModal from "./DeleteAlbumModal";
-import AddFilesToAlbumModal from "./AddFilesToAlbumModal";
+import { processTimelineData } from "../../utils/timelineUtils";
+import ImageViewer from "../viewers/ImageViewer";
+import VideoViewer from "../viewers/VideoViewer";
+import RawImageViewer from "../viewers/RawImageViewer";
+import AddToAlbumModal from "../modals/AddToAlbumModal";
+import { api } from "../../api";
+import formatTime from "../../utils/timeUtils";
+import { formatFileSize } from "../../utils/fileUtils";
+import FavoriteButton from "../buttons/FavoriteButton";
+import ShareButton from "../buttons/ShareButton";
+import DownloadButton from "../buttons/DownloadButton";
+import DeleteButton from "../buttons/DeleteButton";
+import EditAlbumModal from "../modals/EditAlbumModal";
+import DeleteAlbumModal from "../modals/DeleteAlbumModal";
+import AddFilesToAlbumModal from "../modals/AddFilesToAlbumModal";
 import { Download, Edit, Trash2, Plus, Image as ImageIcon, ArrowLeft, Folder, HardDrive, Clock, RefreshCw, MapPin, Video, FileImage, Info, ImagePlus, X, Frame, XCircle, Files, Camera, Image, VideoIcon } from "lucide-react";
 import { format } from "date-fns";
-import SearchBar from "./SearchBar";
+import SearchBar from "../common/SearchBar";
 
 // PhotoCard component for images
 function PhotoCard({ photo, index, onClick, onFavoriteToggle, onRemove, onSetCover, isCurrentCover, onDelete }) {
