+++ frontend/src/components/screens/PhotoGrid.jsx
@@ -1,58 +1,21 @@
 import { motion } from "framer-motion";
-import { Play, Pause } from "lucide-react";
+
 import { useState, useMemo, useRef, useEffect } from "react";
+
 import { Virtuoso } from "react-virtuoso";
-import { processTimelineData } from "../utils/timelineUtils";
-import VideoViewer from "./VideoViewer";
-import AddToAlbumModal from "./AddToAlbumModal";
-import { api } from "../api";
-import formatTime from "../utils/timeUtils";
-import FavoriteButton from "./FavoriteButton";
-import ShareButton from "./ShareButton";
-import DownloadButton from "./DownloadButton";
-import DeleteButton from "./DeleteButton";    
 
-function VideoCard({ video, index, onClick, onFavoriteToggle, onDelete }) {
+import { processTimelineData } from "../../utils/timelineUtils";
+
+import ImageViewer from "../viewers/ImageViewer";
+import AddToAlbumModal from "../modals/AddToAlbumModal";
+
+import FavoriteButton from "../buttons/FavoriteButton";
+import ShareButton from "../buttons/ShareButton";
+import DownloadButton from "../buttons/DownloadButton";
+import DeleteButton from "../buttons/DeleteButton";
+
+function PhotoCard({ photo, index, onClick, onFavoriteToggle, onDelete }) {
   const [isHovered, setIsHovered] = useState(false);
-  const [isPreviewPlaying, setIsPreviewPlaying] = useState(false);
-  const previewRef = useRef(null);
-  const previewTimeoutRef = useRef(null);
-
-  const handleMouseEnter = () => {
-    setIsHovered(true);
-    // Delay preview playback slightly to avoid loading on quick hovers
-    previewTimeoutRef.current = setTimeout(() => {
-      if (previewRef.current) {
-        previewRef.current.play().catch(e => console.log("Preview play failed:", e));
-        setIsPreviewPlaying(true);
-      }
-    }, 300);
-  };
-
-  const handleMouseLeave = () => {
-    setIsHovered(false);
-    setIsPreviewPlaying(false);
-    
-    // Clear the timeout to prevent preview from starting after mouse leaves
-    if (previewTimeoutRef.current) {
-      clearTimeout(previewTimeoutRef.current);
-      previewTimeoutRef.current = null;
-    }
-    
-    if (previewRef.current) {
-      previewRef.current.pause();
-      previewRef.current.currentTime = 0;
-    }
-  };
-
-  // Cleanup timeout on unmount
-  useEffect(() => {
-    return () => {
-      if (previewTimeoutRef.current) {
-        clearTimeout(previewTimeoutRef.current);
-      }
-    };
-  }, []);
 
   return (
     <motion.div
@@ -61,92 +24,75 @@ function VideoCard({ video, index, onClick, onFavoriteToggle, onDelete }) {
       viewport={{ once: true, margin: "50px" }}
       transition={{
         duration: 0.5,
+
         delay: index * 0.05,
+
         type: "spring",
+
         stiffness: 100,
+
         damping: 12,
       }}
       whileHover={{ scale: 1.02, zIndex: 10 }}
-      onHoverStart={handleMouseEnter}
-      onHoverEnd={handleMouseLeave}
+      onHoverStart={() => setIsHovered(true)}
+      onHoverEnd={() => setIsHovered(false)}
       onClick={onClick}
       className="relative group cursor-pointer h-full w-full"
     >
       <div
         className={`
+
 relative overflow-hidden rounded-2xl h-64 w-full
+
 transition-all duration-500
+
 ${
   isHovered
     ? "shadow-glow-cyan border-2 border-purple-400/50 dark:border-cyan-400/40"
     : "shadow-xl border-2 border-slate-200/40 dark:border-white/10"
 }
+
 `}
       >
         {/* Floating Favorite Button - Top Right (when favorited) */}
-        {video.is_favorite && (
+        {photo.is_favorite && (
           <div className="absolute top-3 right-3 z-10">
             <FavoriteButton 
-              id={video.id}
-              type="video"
-              initialFavorite={video.is_favorite}
+              id={photo.id}
+              type="image"
+              initialFavorite={photo.is_favorite}
               size="small"
               onToggle={onFavoriteToggle}
             />
           </div>
         )}
 
-        {/* Video Thumbnail */}
+        {/* Image */}
+
         <img
-          src={api.getVideoThumbnailUrl(video.id)}
-          alt={video.filename}
-          className={`w-full h-full object-cover transition-opacity duration-300 ${isPreviewPlaying ? 'opacity-0' : 'opacity-100'}`}
+          src={photo.thumbnail_url}
+          alt={photo.filename}
+          className="w-full h-full object-cover"
           loading="eager"
           decoding="async"
         />
 
-        {/* Preview Video (plays on hover) */}
-        <video
-          ref={previewRef}
-          src={api.getVideoPreviewUrl(video.id)}
-          className={`absolute inset-0 w-full h-full object-cover transition-opacity duration-300 ${isPreviewPlaying ? 'opacity-100' : 'opacity-0'}`}
-          loop
-          muted
-          playsInline
-          preload="none"
-        />
-
-        {/* Play Icon Overlay */}
-        <div className="absolute inset-0 flex items-center justify-center pointer-events-none">
-          <motion.div
-            initial={{ scale: 0.8, opacity: 0.6 }}
-            animate={{ 
-              scale: isHovered ? (isPreviewPlaying ? 0.9 : 1.1) : 1,
-              opacity: isPreviewPlaying ? 0.3 : (isHovered ? 1 : 0.8)
-            }}
-            transition={{ duration: 0.3 }}
-            className="w-16 h-16 rounded-full bg-black/60 backdrop-blur-sm 
-                     flex items-center justify-center border-2 border-white/30
-                     shadow-2xl"
-          >
-            {isPreviewPlaying ? (
-              <Pause className="w-8 h-8 text-white" fill="white" />
-            ) : (
-              <Play className="w-8 h-8 text-white ml-1" fill="white" />
-            )}
-          </motion.div>
-        </div>
-
         {/* Gradient Overlay */}
+
         <div
           className={`
+
 absolute inset-0 bg-gradient-to-t from-black/80 via-black/20 to-transparent
+
 transition-opacity duration-500
+
 ${isHovered ? "opacity-100" : "opacity-0"}
+
 `}
         />
 
         {/* Content Overlay */}
+
         <motion.div
           initial={{ opacity: 0, y: 20 }}
           animate={{ opacity: isHovered ? 1 : 0, y: isHovered ? 0 : 20 }}
@@ -156,9 +102,10 @@ ${isHovered ? "opacity-100" : "opacity-0"}
         >
           <div className="space-y-1">
             {/* Dynamically render location only if available */}
-            {(video.city || video.state || video.country) && (
+
+            {(photo.city || photo.state || photo.country) && (
               <p className="text-white font-semibold text-sm tracking-wide drop-shadow-lg">
-                {[video.city, video.state, video.country]
+                {[photo.city, photo.state, photo.country]
                   .filter(Boolean)
                   .join(", ")}
               </p>
@@ -166,155 +113,198 @@ ${isHovered ? "opacity-100" : "opacity-0"}
 
             {/* Camera metadata if available */}
 
-            {video.metadata?.camera_model && (
+            {photo.metadata?.camera_model && (
               <p className="text-purple-300 dark:text-cyan-300 text-xs font-bold flex items-center gap-1">
-                {video.metadata.camera_model}
-              </p>
-            )}
-
-            {/* Duration if available */}
-            {(video.duration) && (
-              <p className="text-purple-300 dark:text-cyan-300 text-xs font-bold flex items-center gap-2">
-                {video.duration && (
-                  <span>{formatTime(video.duration)}</span>
-                )}
+                {photo.metadata.camera_model}
               </p>
             )}
           </div>
 
           {/* Action Buttons */}
+
           <div className="flex items-center gap-2 mt-3">
             <FavoriteButton 
-              id={video.id}
-              type="video"
-              initialFavorite={video.is_favorite}
+              id={photo.id}
+              type="image"
+              initialFavorite={photo.is_favorite}
               size="small"
               onToggle={onFavoriteToggle}
             />
 
             <ShareButton 
-              id={video.id}
-              type="video"
+              id={photo.id}
+              type="image"
               size="small"
             />
 
             <DownloadButton 
-              id={video.id}
-              type="video"
+              id={photo.id}
+              type="image"
               size="small"
             />
 
             <div className="flex-1" />
 
             <DeleteButton 
-              id={video.id}
-              type="video"
+              id={photo.id}
+              type="image"
               size="small"
               onSuccess={onDelete}
             />
           </div>
         </motion.div>
+
+        {/* Glow Effect on Hover */}
+
+        {/* <motion.div
+
+animate={{
+
+opacity: isHovered ? [0.3, 0.6, 0.3] : 0,
+
+}}
+
+transition={{
+
+duration: glowDuration,
+
+repeat: Infinity,
+
+ease: 'easeInOut',
+
+delay: glowDelay,
+
+}}
+
+className="absolute inset-0 bg-gradient-to-br
+
+from-purple-400/20 via-transparent to-indigo-400/20
+
+dark:from-cyan-500/20 dark:via-transparent dark:to-teal-500/20
+
+pointer-events-none"
+
+/> */}
       </div>
     </motion.div>
   );
 }
 
-export default function VideoGrid({
-  videos = [],
+export default function PhotoGrid({
+  photos = [],
   loading = false,
   searchQuery = "",
   onLoadMore,
-  hasMore = true,
+  hasMore=true,
   totalCount,
   statusCode,
   onFavoriteToggle,
   onLocationUpdate,
   onDelete
 }) {
-  const [selectedVideo, setSelectedVideo] = useState(null);
+  const [selectedPhoto, setSelectedPhoto] = useState(null);
   const [isAddToAlbumModalOpen, setIsAddToAlbumModalOpen] = useState(false);
+
   const virtuosoRef = useRef(null);
 
-  const sortedVideos = useMemo(() => {
-    return [...videos].sort((a, b) => {
+  const sortedPhotos = useMemo(() => {
+    return [...photos].sort((a, b) => {
       const dateA = new Date(a.date || a.capture_date || 0);
       const dateB = new Date(b.date || b.capture_date || 0);
       return dateB - dateA; // Descending (Newest first)
     });
-  }, [videos]);
+  }, [photos]);
 
-  // Process videos into timeline rows (Year > Month > Video rows)
-  const timelineRows = useMemo(() => processTimelineData(sortedVideos, 5), [sortedVideos]);
+  // Process photos into timeline rows (Year > Month > Photo rows)
 
-  const getSortedIndex = (video) => sortedVideos.findIndex((v) => v.id === video.id);
+  const timelineRows = useMemo(() => processTimelineData(sortedPhotos, 5), [sortedPhotos]);
+
+  const getSortedIndex = (photo) => sortedPhotos.findIndex((p) => p.id === photo.id);
 
   useEffect(() => {
-    if (selectedVideo && virtuosoRef.current) {
-      const rowIndex = timelineRows.findIndex(row => 
-        row.type === 'photos' && row.items.some(v => v.id === selectedVideo.id)
-      );
+    if (selectedPhoto && virtuosoRef.current) {
+        // A. Find which "Row" contains this photo
+        const rowIndex = timelineRows.findIndex(row => 
+            row.type === 'photos' && row.items.some(p => p.id === selectedPhoto.id)
+        );
 
-      if (rowIndex !== -1) {
-        virtuosoRef.current.scrollToIndex({
-          index: rowIndex,
-          align: 'center',
-          behavior: 'auto'
-        });
-      }
+        console.log("Syncing Scroll:", { photoId: selectedPhoto.id, rowIndex }); // <--- DEBUG LOG
+
+        // B. Scroll the background to that row (centered)
+        if (rowIndex !== -1) {
+            virtuosoRef.current.scrollToIndex({
+                index: rowIndex,
+                align: 'center',
+                behavior: 'auto' // Instant scroll, use 'smooth' if you prefer animation
+            });
+        }
     }
-  }, [selectedVideo, timelineRows]);
+  }, [selectedPhoto, timelineRows]); // Run whenever photo changes
 
-  const handleVideoClick = (video) => {
-    setSelectedVideo(video);
+  const handlePhotoClick = (photo) => {
+    setSelectedPhoto(photo);
   };
 
   const handleClose = () => {
-    setSelectedVideo(null);
+    setSelectedPhoto(null);
   };
 
   const handleNext = () => {
-    const currentIdx = getSortedIndex(selectedVideo);
-    if (currentIdx !== -1 && currentIdx < sortedVideos.length - 1) {
-      setSelectedVideo(sortedVideos[currentIdx + 1]);
+    const currentIdx = getSortedIndex(selectedPhoto);
+
+    if (currentIdx !== -1 && currentIdx < sortedPhotos.length - 1) {
+      setSelectedPhoto(sortedPhotos[currentIdx + 1]);
     }
   };
 
   const handlePrev = () => {
-    const currentIdx = getSortedIndex(selectedVideo);
+    const currentIdx = getSortedIndex(selectedPhoto);
+
     if (currentIdx > 0) {
-      setSelectedVideo(sortedVideos[currentIdx - 1]);
+      setSelectedPhoto(sortedPhotos[currentIdx - 1]);
     }
   };
 
-  // Format search query
+  // Format search query: capitalize first letter after periods, show first 4 words
+
   const formatSearchQuery = (query) => {
     if (!query) return "";
+
+    // Capitalize first letter of sentences (after periods and at start)
+
     const formatted = query
       .toLowerCase()
       .replace(/(^|\. )(\w)/g, (match) => match.toUpperCase());
+
+    // Get first 4 words
+
     const words = formatted.split(" ");
+
     if (words.length > 4) {
       return words.slice(0, 4).join(" ") + "...";
     }
+
     return formatted;
   };
 
   const Footer = () => {
     return loading && hasMore ? (
       <div className="py-8 flex justify-center w-full">
-        <div className="w-8 h-8 border-4 border-purple-600/30 border-t-purple-600 rounded-full animate-spin" />
+         <div className="w-8 h-8 border-4 border-purple-600/30 border-t-purple-600 rounded-full animate-spin" />
       </div>
     ) : null;
   };
 
   // --- Loading State ---
-  if (loading && videos.length === 0) {
+
+  if (loading && photos.length === 0) {
     return (
       <div className="min-h-screen pt-32 pb-16 px-8 pl-[calc(240px+6rem)] flex items-center justify-center">
         <div className="text-center">
           <div className="w-16 h-16 border-4 border-purple-600/30 dark:border-cyan-400/30 border-t-purple-600 dark:border-t-cyan-400 rounded-full animate-spin mx-auto mb-4" />
+
           <p className="text-slate-600 dark:text-white/50 text-lg">
-            Loading your videos...
+            Loading your memories...
           </p>
         </div>
       </div>
@@ -329,6 +319,7 @@ export default function VideoGrid({
           <p className="text-slate-600 dark:text-white/50 text-lg">
             Haven Vault is not configured
           </p>
+
           <p className="text-slate-400 dark:text-white/30 text-sm mt-2">
             Please configure Haven Vault to get started!
           </p>
@@ -336,22 +327,23 @@ export default function VideoGrid({
       </div>
     );
   }
-
   // --- Empty State ---
-  if (!loading && (!videos || videos.length === 0)) {
+
+  else if (!loading && (!photos || photos.length === 0)) {
     return (
       <div className="min-h-screen pt-32 pb-16 px-8 pl-[calc(240px+6rem)] flex items-center justify-center">
         <div className="text-center">
           <p className="text-slate-600 dark:text-white/50 text-lg">
-            No {searchQuery ? `"${formatSearchQuery(searchQuery)}"` : ""} videos found in Haven Vault
+            No {searchQuery ? `"${formatSearchQuery(searchQuery)}"` : ""} photos found in Haven Vault
           </p>
+
           {searchQuery ? (
             <p className="text-slate-400 dark:text-white/30 text-sm mt-2">
               Try searching for something else!
             </p>
           ) : (
             <p className="text-slate-400 dark:text-white/30 text-sm mt-2">
-              Upload some videos to get started!
+              Upload some photos to get started!
             </p>
           )}
         </div>
@@ -360,9 +352,11 @@ export default function VideoGrid({
   }
 
   // --- Main Timeline Render ---
+
   return (
     <div className="min-h-screen pt-32 pb-16 px-8 pl-[calc(240px+6rem)]">
       {/* Header Stats */}
+
       <motion.div
         initial={{ opacity: 0, y: 20 }}
         animate={{ opacity: 1, y: 0 }}
@@ -372,24 +366,29 @@ export default function VideoGrid({
         <div>
           <h1
             className="text-4xl font-bold bg-gradient-to-r
+
 from-purple-600 via-indigo-600 to-violet-600
+
 dark:from-white dark:via-cyan-100 dark:to-teal-100
+
 bg-clip-text text-transparent mb-2"
           >
             {searchQuery
-              ? `Searching: "${formatSearchQuery(searchQuery)}" in Your Videos`
-              : "Your Videos"}
+              ? `Searching: "${formatSearchQuery(searchQuery)}" in Your Photos`
+              : "Your Photos"}
           </h1>
+
           <p className="text-slate-600 dark:text-white/50 text-lg">
             <span className="font-semibold text-purple-600 dark:text-cyan-400">
               {totalCount}
             </span>{" "}
-            {searchQuery ? "search results" : "videos"} in Haven Vault
+            {searchQuery ? "search results" : "photos"} in Haven Vault
           </p>
         </div>
       </motion.div>
 
       {/* VIRTUALIZED TIMELINE */}
+
       <Virtuoso
         useWindowScroll
         ref={virtuosoRef}
@@ -399,6 +398,7 @@ bg-clip-text text-transparent mb-2"
         components={{ Footer }}
         itemContent={(index, row) => {
           // 1. Year Header
+
           if (row.type === "year") {
             return (
               <div className="pt-8 pb-4">
@@ -410,33 +410,38 @@ bg-clip-text text-transparent mb-2"
           }
 
           // 2. Month Header
+
           if (row.type === "month") {
             return (
               <div className="pt-0 pb-3 sticky top-[100px] z-20 backdrop-blur-md w-full">
                 <h3 className="text-lg font-medium text-purple-600 dark:text-cyan-400 flex items-center gap-2">
                   <div className="w-2 h-2 rounded-full bg-purple-600 dark:bg-cyan-400 shadow-glow-cyan" />
+
                   {row.label}
                 </h3>
               </div>
             );
           }
 
-          // 3. Video Grid Row
+          // 3. Photo Grid Row
+
           if (row.type === "photos") {
             return (
               <div className="flex gap-4 mb-4 mt-4">
-                {row.items.map((video, i) => (
-                  <div key={video.id} className="flex-1 min-w-0">
-                    <VideoCard
-                      video={video}
+                {row.items.map((photo, i) => (
+                  <div key={photo.id} className="flex-1 min-w-0">
+                    <PhotoCard
+                      photo={photo}
                       index={i}
-                      onClick={() => handleVideoClick(video)}
+                      onClick={() => handlePhotoClick(photo)}
                       onFavoriteToggle={onFavoriteToggle}
                       onDelete={onDelete}
                     />
                   </div>
                 ))}
+
                 {/* Spacers for incomplete rows */}
+
                 {[...Array(5 - row.items.length)].map((_, i) => (
                   <div key={`empty-${i}`} className="flex-1" />
                 ))}
@@ -448,15 +453,16 @@ bg-clip-text text-transparent mb-2"
         }}
       />
 
-      {/* Video Viewer */}
-      {selectedVideo && (
-        <VideoViewer
-          video={selectedVideo}
+      {/* Image Viewer */}
+
+      {selectedPhoto && (
+        <ImageViewer
+          photo={selectedPhoto}
           onClose={handleClose}
           onNext={handleNext}
           onPrev={handlePrev}
-          currentIndex={getSortedIndex(selectedVideo)}
-          totalVideos={totalCount}
+          currentIndex={getSortedIndex(selectedPhoto)}
+          totalPhotos={totalCount}
           onFavoriteToggle={onFavoriteToggle}
           onLocationUpdate={onLocationUpdate}
           onDelete={onDelete}
@@ -466,16 +472,15 @@ bg-clip-text text-transparent mb-2"
       )}
 
       {/* AddToAlbumModal */}
-      {selectedVideo && (
+      {selectedPhoto && (
         <AddToAlbumModal
           isOpen={isAddToAlbumModalOpen}
           onClose={() => setIsAddToAlbumModalOpen(false)}
-          fileId={selectedVideo.id}
-          fileType="video"
-          fileName={selectedVideo.filename}
+          fileId={selectedPhoto.id}
+          fileType="image"
+          fileName={selectedPhoto.filename}
         />
       )}
     </div>
   );
 }
-
