+++ frontend/src/components/screens/VideoGrid.jsx
@@ -1,23 +1,58 @@
 import { motion } from "framer-motion";
-
+import { Play, Pause } from "lucide-react";
 import { useState, useMemo, useRef, useEffect } from "react";
-
 import { Virtuoso } from "react-virtuoso";
+import { processTimelineData } from "../../utils/timelineUtils";
+import VideoViewer from "../viewers/VideoViewer";
+import AddToAlbumModal from "../modals/AddToAlbumModal";
+import { api } from "../../api";
+import formatTime from "../../utils/timeUtils";
+import FavoriteButton from "../buttons/FavoriteButton";
+import ShareButton from "../buttons/ShareButton";
+import DownloadButton from "../buttons/DownloadButton";
+import DeleteButton from "../buttons/DeleteButton";    
 
-import { processTimelineData } from "../utils/timelineUtils";
-
-import RawImageViewer from "./RawImageViewer";
-import AddToAlbumModal from "./AddToAlbumModal";
-
-import { api } from "../api";
-
-import FavoriteButton from "./FavoriteButton";
-import ShareButton from "./ShareButton";
-import DownloadButton from "./DownloadButton";
-import DeleteButton from "./DeleteButton";
-
-function RawImageCard({ rawImage, index, onClick, onFavoriteToggle, onDelete }) {
+function VideoCard({ video, index, onClick, onFavoriteToggle, onDelete }) {
   const [isHovered, setIsHovered] = useState(false);
+  const [isPreviewPlaying, setIsPreviewPlaying] = useState(false);
+  const previewRef = useRef(null);
+  const previewTimeoutRef = useRef(null);
+
+  const handleMouseEnter = () => {
+    setIsHovered(true);
+    // Delay preview playback slightly to avoid loading on quick hovers
+    previewTimeoutRef.current = setTimeout(() => {
+      if (previewRef.current) {
+        previewRef.current.play().catch(e => console.log("Preview play failed:", e));
+        setIsPreviewPlaying(true);
+      }
+    }, 300);
+  };
+
+  const handleMouseLeave = () => {
+    setIsHovered(false);
+    setIsPreviewPlaying(false);
+    
+    // Clear the timeout to prevent preview from starting after mouse leaves
+    if (previewTimeoutRef.current) {
+      clearTimeout(previewTimeoutRef.current);
+      previewTimeoutRef.current = null;
+    }
+    
+    if (previewRef.current) {
+      previewRef.current.pause();
+      previewRef.current.currentTime = 0;
+    }
+  };
+
+  // Cleanup timeout on unmount
+  useEffect(() => {
+    return () => {
+      if (previewTimeoutRef.current) {
+        clearTimeout(previewTimeoutRef.current);
+      }
+    };
+  }, []);
 
   return (
     <motion.div
@@ -26,75 +61,92 @@ function RawImageCard({ rawImage, index, onClick, onFavoriteToggle, onDelete }) 
       viewport={{ once: true, margin: "50px" }}
       transition={{
         duration: 0.5,
-
         delay: index * 0.05,
-
         type: "spring",
-
         stiffness: 100,
-
         damping: 12,
       }}
       whileHover={{ scale: 1.02, zIndex: 10 }}
-      onHoverStart={() => setIsHovered(true)}
-      onHoverEnd={() => setIsHovered(false)}
+      onHoverStart={handleMouseEnter}
+      onHoverEnd={handleMouseLeave}
       onClick={onClick}
       className="relative group cursor-pointer h-full w-full"
     >
       <div
         className={`
-
 relative overflow-hidden rounded-2xl h-64 w-full
-
 transition-all duration-500
-
 ${
   isHovered
     ? "shadow-glow-cyan border-2 border-purple-400/50 dark:border-cyan-400/40"
     : "shadow-xl border-2 border-slate-200/40 dark:border-white/10"
 }
-
 `}
       >
         {/* Floating Favorite Button - Top Right (when favorited) */}
-        {rawImage.is_favorite && (
+        {video.is_favorite && (
           <div className="absolute top-3 right-3 z-10">
             <FavoriteButton 
-              id={rawImage.id}
-              type="raw"
-              initialFavorite={rawImage.is_favorite}
+              id={video.id}
+              type="video"
+              initialFavorite={video.is_favorite}
               size="small"
               onToggle={onFavoriteToggle}
             />
           </div>
         )}
 
-        {/* Image Thumbnail */}
-
+        {/* Video Thumbnail */}
         <img
-          src={api.getRawThumbnailUrl(rawImage.id)}
-          alt={rawImage.filename}
-          className="w-full h-full object-cover"
+          src={api.getVideoThumbnailUrl(video.id)}
+          alt={video.filename}
+          className={`w-full h-full object-cover transition-opacity duration-300 ${isPreviewPlaying ? 'opacity-0' : 'opacity-100'}`}
           loading="eager"
           decoding="async"
         />
 
-        {/* Gradient Overlay */}
+        {/* Preview Video (plays on hover) */}
+        <video
+          ref={previewRef}
+          src={api.getVideoPreviewUrl(video.id)}
+          className={`absolute inset-0 w-full h-full object-cover transition-opacity duration-300 ${isPreviewPlaying ? 'opacity-100' : 'opacity-0'}`}
+          loop
+          muted
+          playsInline
+          preload="none"
+        />
 
+        {/* Play Icon Overlay */}
+        <div className="absolute inset-0 flex items-center justify-center pointer-events-none">
+          <motion.div
+            initial={{ scale: 0.8, opacity: 0.6 }}
+            animate={{ 
+              scale: isHovered ? (isPreviewPlaying ? 0.9 : 1.1) : 1,
+              opacity: isPreviewPlaying ? 0.3 : (isHovered ? 1 : 0.8)
+            }}
+            transition={{ duration: 0.3 }}
+            className="w-16 h-16 rounded-full bg-black/60 backdrop-blur-sm 
+                     flex items-center justify-center border-2 border-white/30
+                     shadow-2xl"
+          >
+            {isPreviewPlaying ? (
+              <Pause className="w-8 h-8 text-white" fill="white" />
+            ) : (
+              <Play className="w-8 h-8 text-white ml-1" fill="white" />
+            )}
+          </motion.div>
+        </div>
+
+        {/* Gradient Overlay */}
         <div
           className={`
-
 absolute inset-0 bg-gradient-to-t from-black/80 via-black/20 to-transparent
-
 transition-opacity duration-500
-
 ${isHovered ? "opacity-100" : "opacity-0"}
-
 `}
         />
 
         {/* Content Overlay */}
-
         <motion.div
           initial={{ opacity: 0, y: 20 }}
           animate={{ opacity: isHovered ? 1 : 0, y: isHovered ? 0 : 20 }}
@@ -104,10 +156,9 @@ ${isHovered ? "opacity-100" : "opacity-0"}
         >
           <div className="space-y-1">
             {/* Dynamically render location only if available */}
-
-            {(rawImage.city || rawImage.state || rawImage.country) && (
+            {(video.city || video.state || video.country) && (
               <p className="text-white font-semibold text-sm tracking-wide drop-shadow-lg">
-                {[rawImage.city, rawImage.state, rawImage.country]
+                {[video.city, video.state, video.country]
                   .filter(Boolean)
                   .join(", ")}
               </p>
@@ -115,48 +166,49 @@ ${isHovered ? "opacity-100" : "opacity-0"}
 
             {/* Camera metadata if available */}
 
-            {rawImage.metadata?.camera_model && (
+            {video.metadata?.camera_model && (
               <p className="text-purple-300 dark:text-cyan-300 text-xs font-bold flex items-center gap-1">
-                {rawImage.metadata.camera_model}
+                {video.metadata.camera_model}
               </p>
             )}
 
-            {/* Lens info if available */}
-            {rawImage.metadata?.lens_model && (
-              <p className="text-purple-200 dark:text-cyan-200 text-xs flex items-center gap-1">
-                {rawImage.metadata.lens_model}
+            {/* Duration if available */}
+            {(video.duration) && (
+              <p className="text-purple-300 dark:text-cyan-300 text-xs font-bold flex items-center gap-2">
+                {video.duration && (
+                  <span>{formatTime(video.duration)}</span>
+                )}
               </p>
             )}
           </div>
 
           {/* Action Buttons */}
-
           <div className="flex items-center gap-2 mt-3">
             <FavoriteButton 
-              id={rawImage.id}
-              type="raw"
-              initialFavorite={rawImage.is_favorite}
+              id={video.id}
+              type="video"
+              initialFavorite={video.is_favorite}
               size="small"
               onToggle={onFavoriteToggle}
             />
 
             <ShareButton 
-              id={rawImage.id}
-              type="raw"
+              id={video.id}
+              type="video"
               size="small"
             />
 
             <DownloadButton 
-              id={rawImage.id}
-              type="raw"
+              id={video.id}
+              type="video"
               size="small"
             />
 
             <div className="flex-1" />
 
             <DeleteButton 
-              id={rawImage.id}
-              type="raw"
+              id={video.id}
+              type="video"
               size="small"
               onSuccess={onDelete}
             />
@@ -167,119 +219,102 @@ ${isHovered ? "opacity-100" : "opacity-0"}
   );
 }
 
-export default function RawImageGrid({
-  rawImages = [],
+export default function VideoGrid({
+  videos = [],
   loading = false,
   searchQuery = "",
   onLoadMore,
-  hasMore=true,
+  hasMore = true,
   totalCount,
   statusCode,
   onFavoriteToggle,
   onLocationUpdate,
   onDelete
 }) {
-  const [selectedRawImage, setSelectedRawImage] = useState(null);
+  const [selectedVideo, setSelectedVideo] = useState(null);
   const [isAddToAlbumModalOpen, setIsAddToAlbumModalOpen] = useState(false);
-
   const virtuosoRef = useRef(null);
 
-  const sortedRawImages = useMemo(() => {
-    return [...rawImages].sort((a, b) => {
+  const sortedVideos = useMemo(() => {
+    return [...videos].sort((a, b) => {
       const dateA = new Date(a.date || a.capture_date || 0);
       const dateB = new Date(b.date || b.capture_date || 0);
       return dateB - dateA; // Descending (Newest first)
     });
-  }, [rawImages]);
+  }, [videos]);
 
-  // Process raw images into timeline rows (Year > Month > Photo rows)
+  // Process videos into timeline rows (Year > Month > Video rows)
+  const timelineRows = useMemo(() => processTimelineData(sortedVideos, 5), [sortedVideos]);
 
-  const timelineRows = useMemo(() => processTimelineData(sortedRawImages, 5), [sortedRawImages]);
-
-  const getSortedIndex = (rawImage) => sortedRawImages.findIndex((r) => r.id === rawImage.id);
+  const getSortedIndex = (video) => sortedVideos.findIndex((v) => v.id === video.id);
 
   useEffect(() => {
-    if (selectedRawImage && virtuosoRef.current) {
-        // A. Find which "Row" contains this raw image
-        const rowIndex = timelineRows.findIndex(row => 
-            row.type === 'photos' && row.items.some(r => r.id === selectedRawImage.id)
-        );
+    if (selectedVideo && virtuosoRef.current) {
+      const rowIndex = timelineRows.findIndex(row => 
+        row.type === 'photos' && row.items.some(v => v.id === selectedVideo.id)
+      );
 
-        // B. Scroll the background to that row (centered)
-        if (rowIndex !== -1) {
-            virtuosoRef.current.scrollToIndex({
-                index: rowIndex,
-                align: 'center',
-                behavior: 'auto' // Instant scroll, use 'smooth' if you prefer animation
-            });
-        }
+      if (rowIndex !== -1) {
+        virtuosoRef.current.scrollToIndex({
+          index: rowIndex,
+          align: 'center',
+          behavior: 'auto'
+        });
+      }
     }
-  }, [selectedRawImage, timelineRows]); // Run whenever raw image changes
+  }, [selectedVideo, timelineRows]);
 
-  const handleRawImageClick = (rawImage) => {
-    setSelectedRawImage(rawImage);
+  const handleVideoClick = (video) => {
+    setSelectedVideo(video);
   };
 
   const handleClose = () => {
-    setSelectedRawImage(null);
+    setSelectedVideo(null);
   };
 
   const handleNext = () => {
-    const currentIdx = getSortedIndex(selectedRawImage);
-
-    if (currentIdx !== -1 && currentIdx < sortedRawImages.length - 1) {
-      setSelectedRawImage(sortedRawImages[currentIdx + 1]);
+    const currentIdx = getSortedIndex(selectedVideo);
+    if (currentIdx !== -1 && currentIdx < sortedVideos.length - 1) {
+      setSelectedVideo(sortedVideos[currentIdx + 1]);
     }
   };
 
   const handlePrev = () => {
-    const currentIdx = getSortedIndex(selectedRawImage);
-
+    const currentIdx = getSortedIndex(selectedVideo);
     if (currentIdx > 0) {
-      setSelectedRawImage(sortedRawImages[currentIdx - 1]);
+      setSelectedVideo(sortedVideos[currentIdx - 1]);
     }
   };
 
-  // Format search query: capitalize first letter after periods, show first 4 words
-
+  // Format search query
   const formatSearchQuery = (query) => {
     if (!query) return "";
-
-    // Capitalize first letter of sentences (after periods and at start)
-
     const formatted = query
       .toLowerCase()
       .replace(/(^|\. )(\w)/g, (match) => match.toUpperCase());
-
-    // Get first 4 words
-
     const words = formatted.split(" ");
-
     if (words.length > 4) {
       return words.slice(0, 4).join(" ") + "...";
     }
-
     return formatted;
   };
 
   const Footer = () => {
     return loading && hasMore ? (
       <div className="py-8 flex justify-center w-full">
-         <div className="w-8 h-8 border-4 border-purple-600/30 border-t-purple-600 rounded-full animate-spin" />
+        <div className="w-8 h-8 border-4 border-purple-600/30 border-t-purple-600 rounded-full animate-spin" />
       </div>
     ) : null;
   };
 
   // --- Loading State ---
-
-  if (loading && rawImages.length === 0) {
+  if (loading && videos.length === 0) {
     return (
       <div className="min-h-screen pt-32 pb-16 px-8 pl-[calc(240px+6rem)] flex items-center justify-center">
         <div className="text-center">
           <div className="w-16 h-16 border-4 border-purple-600/30 dark:border-cyan-400/30 border-t-purple-600 dark:border-t-cyan-400 rounded-full animate-spin mx-auto mb-4" />
-
           <p className="text-slate-600 dark:text-white/50 text-lg">
-            Loading your RAW files...
+            Loading your videos...
           </p>
         </div>
       </div>
@@ -294,7 +329,6 @@ export default function RawImageGrid({
           <p className="text-slate-600 dark:text-white/50 text-lg">
             Haven Vault is not configured
           </p>
-
           <p className="text-slate-400 dark:text-white/30 text-sm mt-2">
             Please configure Haven Vault to get started!
           </p>
@@ -302,23 +336,23 @@ export default function RawImageGrid({
       </div>
     );
   }
-  // --- Empty State ---
 
-  else if (!loading && (!rawImages || rawImages.length === 0)) {
+  // --- Empty State ---
+  if (!loading && (!videos || videos.length === 0)) {
     return (
       <div className="min-h-screen pt-32 pb-16 px-8 pl-[calc(240px+6rem)] flex items-center justify-center">
         <div className="text-center">
           <p className="text-slate-600 dark:text-white/50 text-lg">
-            No {searchQuery ? `"${formatSearchQuery(searchQuery)}"` : ""} RAW images found in Haven Vault
+            No {searchQuery ? `"${formatSearchQuery(searchQuery)}"` : ""} videos found in Haven Vault
           </p>
           {searchQuery ? (
             <p className="text-slate-400 dark:text-white/30 text-sm mt-2">
               Try searching for something else!
             </p>
           ) : (
-          <p className="text-slate-400 dark:text-white/30 text-sm mt-2">
-            Upload some RAW files to get started!
-          </p>
+            <p className="text-slate-400 dark:text-white/30 text-sm mt-2">
+              Upload some videos to get started!
+            </p>
           )}
         </div>
       </div>
@@ -326,11 +360,9 @@ export default function RawImageGrid({
   }
 
   // --- Main Timeline Render ---
-
   return (
     <div className="min-h-screen pt-32 pb-16 px-8 pl-[calc(240px+6rem)]">
       {/* Header Stats */}
-
       <motion.div
         initial={{ opacity: 0, y: 20 }}
         animate={{ opacity: 1, y: 0 }}
@@ -340,29 +372,24 @@ export default function RawImageGrid({
         <div>
           <h1
             className="text-4xl font-bold bg-gradient-to-r
-
 from-purple-600 via-indigo-600 to-violet-600
-
 dark:from-white dark:via-cyan-100 dark:to-teal-100
-
 bg-clip-text text-transparent mb-2"
           >
             {searchQuery
-              ? `Searching: "${formatSearchQuery(searchQuery)}" in Your RAW Images`
-              : "Your RAW Images"}
+              ? `Searching: "${formatSearchQuery(searchQuery)}" in Your Videos`
+              : "Your Videos"}
           </h1>
-
           <p className="text-slate-600 dark:text-white/50 text-lg">
             <span className="font-semibold text-purple-600 dark:text-cyan-400">
               {totalCount}
             </span>{" "}
-            {searchQuery ? "search results" : "RAW images"} in Haven Vault
+            {searchQuery ? "search results" : "videos"} in Haven Vault
           </p>
         </div>
       </motion.div>
 
       {/* VIRTUALIZED TIMELINE */}
-
       <Virtuoso
         useWindowScroll
         ref={virtuosoRef}
@@ -372,7 +399,6 @@ bg-clip-text text-transparent mb-2"
         components={{ Footer }}
         itemContent={(index, row) => {
           // 1. Year Header
-
           if (row.type === "year") {
             return (
               <div className="pt-8 pb-4">
@@ -384,38 +410,33 @@ bg-clip-text text-transparent mb-2"
           }
 
           // 2. Month Header
-
           if (row.type === "month") {
             return (
               <div className="pt-0 pb-3 sticky top-[100px] z-20 backdrop-blur-md w-full">
                 <h3 className="text-lg font-medium text-purple-600 dark:text-cyan-400 flex items-center gap-2">
                   <div className="w-2 h-2 rounded-full bg-purple-600 dark:bg-cyan-400 shadow-glow-cyan" />
-
                   {row.label}
                 </h3>
               </div>
             );
           }
 
-          // 3. Photo Grid Row
-
+          // 3. Video Grid Row
           if (row.type === "photos") {
             return (
               <div className="flex gap-4 mb-4 mt-4">
-                {row.items.map((rawImage, i) => (
-                  <div key={rawImage.id} className="flex-1 min-w-0">
-                    <RawImageCard
-                      rawImage={rawImage}
+                {row.items.map((video, i) => (
+                  <div key={video.id} className="flex-1 min-w-0">
+                    <VideoCard
+                      video={video}
                       index={i}
-                      onClick={() => handleRawImageClick(rawImage)}
+                      onClick={() => handleVideoClick(video)}
                       onFavoriteToggle={onFavoriteToggle}
                       onDelete={onDelete}
                     />
                   </div>
                 ))}
-
                 {/* Spacers for incomplete rows */}
-
                 {[...Array(5 - row.items.length)].map((_, i) => (
                   <div key={`empty-${i}`} className="flex-1" />
                 ))}
@@ -427,16 +448,15 @@ bg-clip-text text-transparent mb-2"
         }}
       />
 
-      {/* Image Viewer */}
-
-      {selectedRawImage && (
-        <RawImageViewer
-          rawImage={selectedRawImage}
+      {/* Video Viewer */}
+      {selectedVideo && (
+        <VideoViewer
+          video={selectedVideo}
           onClose={handleClose}
           onNext={handleNext}
           onPrev={handlePrev}
-          currentIndex={getSortedIndex(selectedRawImage)}
-          totalRawImages={totalCount}
+          currentIndex={getSortedIndex(selectedVideo)}
+          totalVideos={totalCount}
           onFavoriteToggle={onFavoriteToggle}
           onLocationUpdate={onLocationUpdate}
           onDelete={onDelete}
@@ -446,13 +466,13 @@ bg-clip-text text-transparent mb-2"
       )}
 
       {/* AddToAlbumModal */}
-      {selectedRawImage && (
+      {selectedVideo && (
         <AddToAlbumModal
           isOpen={isAddToAlbumModalOpen}
           onClose={() => setIsAddToAlbumModalOpen(false)}
-          fileId={selectedRawImage.id}
-          fileType="raw"
-          fileName={selectedRawImage.filename}
+          fileId={selectedVideo.id}
+          fileType="video"
+          fileName={selectedVideo.filename}
         />
       )}
     </div>
